Author: Zbigniew Jędrzejewski-Szmek <zbyszek@in.waw.pl> Date:   Tue May 7 15:58:29 2019 +0200
scsi_serial: replace some crazy strncpy() calls by strnlen()

gcc was warning about strncpy() leaving an unterminated string.
In this case, it was correct.

The code was doing strncpy()+strncat()+strlen() essentially to determine
if the strings have expected length. If the length was correct, a buffer
overread was performed (or at least some garbage bytes were used from the
uninitialized part of the buffer). Let's do the length check first and then
only copy stuff if everything agrees.

For some reason the function was called "prepend", when it obviously does
an "append".

systemd-commit: 099c77fd5ff835614dea8dc11c57f6d44f77d9ee